# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Piotr Konkol - Praca in\u017cynierska - Podsumowania z u\u017cyciem LLM",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/research/evaluation-sets": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Evaluation Sets",
        "operationId": "list_evaluation_sets_api_v1_research_evaluation_sets_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvaluationSetListItemResponse"
                  },
                  "type": "array",
                  "title": "Response List Evaluation Sets Api V1 Research Evaluation Sets Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Evaluation Set",
        "operationId": "create_evaluation_set_api_v1_research_evaluation_sets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationSetImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/research/evaluation-sets/{set_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Evaluation Set",
        "operationId": "get_evaluation_set_api_v1_research_evaluation_sets__set_id__get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "research"
        ],
        "summary": "Delete Evaluation Set",
        "operationId": "delete_evaluation_set_api_v1_research_evaluation_sets__set_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetDeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/evaluation-sets/{set_id}/export": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Export Evaluation Set",
        "operationId": "export_evaluation_set_api_v1_research_evaluation_sets__set_id__export_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/evaluation-sets/{set_id}/entries/{entry_id}/input-text": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Evaluation Set Entry Input Text",
        "operationId": "get_evaluation_set_entry_input_text_api_v1_research_evaluation_sets__set_id__entries__entry_id__input_text_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          },
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationSetEntryInputTextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/research/evaluation-sets/{set_id}/runs": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Evaluation Run",
        "operationId": "create_evaluation_run_api_v1_research_evaluation_sets__set_id__runs_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluationRunCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Evaluation Runs",
        "operationId": "list_evaluation_runs_api_v1_research_evaluation_sets__set_id__runs_get",
        "parameters": [
          {
            "name": "set_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Set Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationRunListItemResponse"
                  },
                  "title": "Response List Evaluation Runs Api V1 Research Evaluation Sets  Set Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs/{run_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Evaluation Run",
        "operationId": "get_evaluation_run_api_v1_research_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "research"
        ],
        "summary": "Delete Evaluation Run",
        "operationId": "delete_evaluation_run_api_v1_research_runs__run_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunDeletedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs/{run_id}/entries": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Get Evaluation Run Entries",
        "operationId": "get_evaluation_run_entries_api_v1_research_runs__run_id__entries_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationRunEntriesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/research/runs/{run_id}/deepeval": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Evaluate Run Deepeval",
        "operationId": "evaluate_run_deepeval_api_v1_research_runs__run_id__deepeval_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeepevalQueuedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "EvaluationRunCreateResponse": {
        "properties": {
          "evaluation_run_id": {
            "type": "string",
            "title": "Evaluation Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "evaluation_run_id",
          "status",
          "created_at"
        ],
        "title": "EvaluationRunCreateResponse"
      },
...
      "EvaluationSetCreateResponse": {
        "properties": {
          "evaluation_set_id": {
            "type": "string",
            "title": "Evaluation Set Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "entry_count": {
            "type": "integer",
            "title": "Entry Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "evaluation_set_id",
          "name",
          "language",
          "entry_count",
          "created_at"
        ],
        "title": "EvaluationSetCreateResponse"
      },
      "EvaluationSetDeletedResponse": {
        "properties": {
          "status": {
            "type": "string",
            "const": "deleted",
            "title": "Status"
          },
          "evaluation_set_id": {
            "type": "string",
            "title": "Evaluation Set Id"
          },
          "deleted_runs": {
            "type": "integer",
            "title": "Deleted Runs"
          }
        },
        "type": "object",
        "required": [
          "status",
          "evaluation_set_id",
          "deleted_runs"
        ],
        "title": "EvaluationSetDeletedResponse"
      },
      "EvaluationSetDetailResponse": {
        "properties": {
          "evaluation_set_id": {
            "type": "string",
            "title": "Evaluation Set Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSetEntryResponse"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "evaluation_set_id",
          "name",
          "language",
          "created_at",
          "entries"
        ],
        "title": "EvaluationSetDetailResponse"
      },
...
      "EvaluationSetEntryInputTextResponse": {
        "properties": {
          "entry_id": {
            "type": "string",
            "title": "Entry Id"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          }
        },
        "type": "object",
        "required": [
          "entry_id",
          "input_text"
        ],
        "title": "EvaluationSetEntryInputTextResponse"
      },
...
      "EvaluationSetExportResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSetEntryImport-Output"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "name",
          "language",
          "entries"
        ],
        "title": "EvaluationSetExportResponse",
        "description": "Mirrors EvaluationSetImportRequest so an exported file can be imported back."
      },
      "EvaluationSetImportRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/EvaluationSetEntryImport-Input"
            },
            "type": "array",
            "title": "Entries"
          }
        },
        "type": "object",
        "required": [
          "name",
          "entries"
        ],
        "title": "EvaluationSetImportRequest"
      },
      "EvaluationSetListItemResponse": {
        "properties": {
          "evaluation_set_id": {
            "type": "string",
            "title": "Evaluation Set Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "entry_count": {
            "type": "integer",
            "title": "Entry Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "evaluation_set_id",
          "name",
          "language",
          "entry_count",
          "created_at"
        ],
        "title": "EvaluationSetListItemResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}